% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collect_events}
\alias{collect_events}
\alias{collect_summaries}
\alias{collect_scalars}
\title{Collect data from tfevents records}
\usage{
collect_events(logdir = get_default_logdir())

collect_summaries(logdir = get_default_logdir())

collect_scalars(logdir = get_default_logdir())
}
\value{
A \code{tibble} with the collected events.
}
\description{
Collects all events of a kind in a single data.frame ready for analysis.
}
\section{Functions}{
\itemize{
\item \code{collect_summaries()}: Collect events that contain summary values.

\item \code{collect_scalars()}: Collect event summaries that contain scalar values.

}}
\examples{
temp <- tempfile()
with_logdir(temp, {
  for(i in 1:5) {
    log_event(my_log = runif(1))
  }
})
# collect all events in files, including file description events
collect_events(temp)
# collect summaries in the logdir
collect_summaries(temp)
# collect only scalar events
collect_scalars()

}
\seealso{
iter_events
}
