% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{iter_events}
\alias{iter_events}
\alias{iter_summaries}
\alias{iter_scalars}
\title{Creates an iterator for events in tfevents records}
\usage{
iter_events(logdir = get_default_logdir())

iter_summaries(logdir = get_default_logdir())

iter_scalars(logdir = get_default_logdir())
}
\arguments{
\item{logdir}{The log directory that you want to query events from.}
}
\value{
An iterator that can be used to get events one by one. Returned iterators use
the \link[coro:iterator]{coro::iterator} protocol.
}
\description{
Allows iterating trough events in tfevents records files without necessarily
loading all of them in RAM at once. Uses \link[coro:iterator]{coro::iterator} protocol.
}
\section{Functions}{
\itemize{
\item \code{iter_summaries()}: Creates an iteartor that discard events that don't contain summaries.

\item \code{iter_scalars()}: Creates an iterator that onyl takes scalar summaries.

}}
\examples{
temp <- tempfile()
with_logdir(temp, {
  for(i in 1:5) {
    log_event(my_log = runif(1))
  }
})

# iterate over all events
iter <- iter_events(temp)
iter()
iter()
coro::collect(iter)

# iterate over summaries only
iter <- iter_summaries(temp)
iter()
iter()
coro::collect(iter)

# iterate over events that contain scalar summaries only
iter <- iter_scalars(temp)
iter()
iter()
coro::collect(iter)

}
\seealso{
collect_events
}
