% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorboard.R
\name{tensorboard}
\alias{tensorboard}
\title{Launch TensorBoard}
\usage{
tensorboard(log_dir, host = "127.0.0.1", port = 6060)
}
\arguments{
\item{log_dir}{Path to the log directory.}

\item{host}{Host to bind TensorBoard to. Defaults to \code{"127.0.0.1"}.}

\item{port}{Port to bind TensorBoard to. Defaults to \code{6060}.}
}
\value{
A \link[processx:process]{processx::process} object (invisibly).
}
\description{
Starts a TensorBoard server to visualize logged events.
}
