% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{value}
\alias{value}
\title{Extracts the value of a summary value}
\usage{
value(x, ...)
}
\arguments{
\item{x}{A \code{tfevents_summary_values} object.}

\item{...}{Currently unused. To allow future extension.}
}
\value{
Depending on the type of the summary it returns an image, audio, text or
scalar.
}
\description{
Summaries are complicated objects because they reflect the Protobuf object
structure that are serialized in the tfevents records files. This function
allows one to easily query vaues from summaries and will dispatch to the
correct way to extract images, audio, text, etc from summary values.
}
\examples{
temp <- tempfile()
with_logdir(temp, {
  for(i in 1:5) {
    log_event(my_log = runif(1))
  }
})

# iterate over all events
summary <- collect_events(temp, n = 1, type = "summary")
value(summary$summary)

}
