% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{write_event}
\alias{write_event}
\title{Write event data into disk}
\usage{
write_event(data, name, step)
}
\arguments{
\item{data}{Event data.}

\item{name}{Name/tag of the event data.}

\item{step}{Global step number}
}
\description{
Write event data into disk
}
\keyword{internal}
