% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{summary_image}
\alias{summary_image}
\alias{summary_image.array}
\alias{summary_image.raw}
\title{Creates a image summary}
\usage{
summary_image(img, ..., metadata = NULL)

\method{summary_image}{array}(img, ..., metadata = NULL)

\method{summary_image}{raw}(img, ..., width, height, colorspace, metadata = NULL)
}
\arguments{
\item{img}{An object that can be converted to an image.}

\item{...}{Currently unused.}

\item{metadata}{A \code{metadata} object, as created with \code{\link[=summary_metadata]{summary_metadata()}}. In
most cases you don't need to change the default.}

\item{width}{Width of the image.}

\item{height}{Height of the image.}

\item{colorspace}{Valid colorspace values are
\code{1 - grayscale},
\code{2 - grayscale + alpha},
\code{3 - RGB},
\code{4 - RGBA},
\code{5 - DIGITAL_YUV},
\code{6 - BGRA}}
}
\description{
Creates a image summary
}
\section{Methods (by class)}{
\itemize{
\item \code{summary_image(array)}: Creates an image from an R array. The array should be
numeric, with values between 0 and 1.

\item \code{summary_image(raw)}: Creates an image from a png encoded image. Eg, created
with \code{\link[png:writePNG]{png::writePNG()}}. In this case you need to provide \code{width}, \code{height} and
\code{colorspace} arguments.

}}
