% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{summary_metadata}
\alias{summary_metadata}
\title{Summary metadata}
\usage{
summary_metadata(
  plugin_name,
  display_name = NA_character_,
  description = NA_character_,
  ...,
  plugin_content = NA
)
}
\arguments{
\item{plugin_name}{The name of the TensorBoard plugin that might use the summary.}

\item{display_name}{Display name for the summary.}

\item{description}{A description of the summary.}

\item{...}{Currently unused. For future expansion.}

\item{plugin_content}{An optional plugin content. Note that it will only be
used if the C++ function \code{make_plugin_data} is aware of \code{plugin_content}
for the specified plugin name. For advanced use only.}
}
\value{
A \code{summary_metadata} object.
}
\description{
Creates a summary metadata that can be passed to multiple \code{summary_} functions.
}
\examples{
summary <- summary_scalar(1, metadata = summary_metadata("scalars"))

}
